oculus-5.3.0-v59-automation @Commit b885242af1f3876c08565808f084a5e2db45402b
OVRPlugin 1.91.0
Platform SDK 1.76.0
Audio SDK 1.41.0

+++ Engine/Plugins/Online/OnlineSubsystemUtils/Source/OnlineSubsystemUtils/Private/IpConnection.cpp
@@ -128,7 +128,7 @@ void UIpConnection::InitBase(UNetDriver* InDriver, class FSocket* InSocket, cons
 
 	if (CVarNetEnableCongestionControl.GetValueOnAnyThread() > 0)
 	{
-		NetworkCongestionControl.Emplace(CurrentNetSpeed, FNetPacketNotify::SequenceHistoryT::Size);
+		NetworkCongestionControl.Emplace(CurrentNetSpeed, uint32(FNetPacketNotify::SequenceHistoryT::Size));
 	}
 }
 
@@ -551,7 +551,7 @@ void UIpConnection::LowLevelSend(void* Data, int32 CountBits, FOutPacketTraits& 
 				// Always flush this profiler data now. Technically this could be incorrect if the send in the task fails,
 				// but this keeps the bookkeeping simpler for now.
 				NETWORK_PROFILER(GNetworkProfiler.FlushOutgoingBunches(this));
-				NETWORK_PROFILER(GNetworkProfiler.TrackSocketSendTo(CurSocket->GetDescription(), DataToSend, CountBytes, NumPacketIdBits, NumBunchBits, NumAckBits, NumPaddingBits, this));
+				NETWORK_PROFILER(GNetworkProfiler.TrackSocketSendTo(CurSocket->GetDescription(), DataToSend, uint16(CountBytes), uint16(NumPacketIdBits), uint16(NumBunchBits), uint16(NumAckBits), uint16(NumPaddingBits), this));
 			}
 			else
 			{
@@ -565,7 +565,7 @@ void UIpConnection::LowLevelSend(void* Data, int32 CountBits, FOutPacketTraits& 
 				{
 					UNCLOCK_CYCLES(Driver->SendCycles);
 					NETWORK_PROFILER(GNetworkProfiler.FlushOutgoingBunches(this));
-					NETWORK_PROFILER(GNetworkProfiler.TrackSocketSendTo(CurSocket->GetDescription(), DataToSend, SendResult.BytesSent, NumPacketIdBits, NumBunchBits, NumAckBits, NumPaddingBits, this));
+					NETWORK_PROFILER(GNetworkProfiler.TrackSocketSendTo(CurSocket->GetDescription(), DataToSend, uint16(SendResult.BytesSent), uint16(NumPacketIdBits), uint16(NumBunchBits), uint16(NumAckBits), uint16(NumPaddingBits), this));
 
 					if (bNotifyOnSuccess)
 					{
@@ -655,7 +655,7 @@ void UIpConnection::HandleSocketSendResult(const FSocketSendResult& Result, ISoc
 			const double Time = Driver->GetElapsedTime();
 			if (SocketError_SendDelayStartTime == 0.0)
 			{
-				UE_LOG(LogNet, Log, TEXT("UIpConnection::HandleSocketSendResult: Socket->SendTo failed with error %i (%s). %s Connection beginning close timeout (Timeout = %d)."),
+				UE_LOG(LogNet, Log, TEXT("UIpConnection::HandleSocketSendResult: Socket->SendTo failed with error %i (%s). %s Connection beginning close timeout (Timeout = %f)."),
 					static_cast<int32>(Result.Error),
 					SocketSubsystem->GetSocketError(Result.Error),
 					*Describe(),
@@ -708,7 +708,7 @@ void UIpConnection::HandleSocketRecvError(class UNetDriver* NetDriver, const FSt
 		const double Time = NetDriver->GetElapsedTime();
 		if (SocketError_RecvDelayStartTime == 0.0)
 		{
-			UE_LOG(LogNet, Log, TEXT("%s. %s Connection beginning close timeout (Timeout = %d)."),
+			UE_LOG(LogNet, Log, TEXT("%s. %s Connection beginning close timeout (Timeout = %f)."),
 				*ErrorString,
 				*Describe(),
 				SocketErrorDisconnectDelay);
